Fix:Доработаны замечания
