+++ src/main/java/com/example/Feline.java
@@ -21,5 +21,4 @@ public class Feline extends Animal implements Predator {
     public int getKittens(int kittensCount) {
         return kittensCount;
     }
-
 }
