+++ src/test/java/FelineTest.java
@@ -1,8 +1,6 @@
 package com.example;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
@@ -14,22 +12,16 @@ class FelineTest {
     @Test
     void testEatMeat() throws Exception {
         List<String> expectedFood = List.of("Животные", "Птицы", "Рыба");
-        assertEquals(expectedFood, feline.eatMeat());
+        assertEquals(expectedFood, feline.eatMeat(),"Кошачьи такое не едят!");
     }
 
     @Test
     void testGetFamily() {
-        assertEquals("Кошачьи", feline.getFamily());
+        assertEquals("Кошачьи", feline.getFamily(),"Неверная семья");
     }
 
     @Test
     void testGetKittensWithoutParameters() {
         assertEquals(1, feline.getKittens());
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = {0, 1, 5, 10})
-    void testGetKittensWithParameters(int kittensCount) {
-        assertEquals(kittensCount, feline.getKittens(kittensCount));
     }
 }
