+++ src/test/java/LionTest.java
@@ -1,7 +1,6 @@
 package com.example;
+
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -18,43 +17,33 @@ class LionTest {
 
     @Test
     void testDoesHaveManeForMale() throws Exception {
-        Lion lion = new Lion("Самец");
+        Lion lion = new Lion("Самец",feline);
         assertTrue(lion.doesHaveMane());
     }
 
     @Test
     void testDoesHaveManeForFemale() throws Exception {
-        Lion lion = new Lion("Самка");
+        Lion lion = new Lion("Самка",feline);
         assertFalse(lion.doesHaveMane());
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"", " ", "Неизвестный", "123"})
-    void testLionConstructorWithInvalidSex(String sex) {
-        Exception exception = assertThrows(Exception.class, () -> new Lion(sex));
-        assertEquals("Используйте допустимые значения пола животного - самей или самка",
-                exception.getMessage());
     }
 
     @Test
     void testGetKittens() throws Exception {
-        Lion lion = new Lion("Самец");
+        Lion lion = new Lion("Самец",feline);
         lion.feline = feline;
         when(feline.getKittens()).thenReturn(1);
 
         assertEquals(1, lion.getKittens());
-        verify(feline, times(1)).getKittens();
     }
 
     @Test
     void testGetFood() throws Exception {
-        Lion lion = new Lion("Самка");
+        Lion lion = new Lion("Самка",feline);
         lion.feline = feline;
         List<String> expectedFood = List.of("Животные", "Птицы", "Рыба");
 
         when(feline.getFood("Хищник")).thenReturn(expectedFood);
 
         assertEquals(expectedFood, lion.getFood());
-        verify(feline, times(1)).getFood("Хищник");
     }
 }
